probability/pyro-code/requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
scipy>=1.7.0
torch>=1.9.0
pyro-ppl>=1.8.0 